infra-flask/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7        
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multi-key-dict==2.0.3
pbr==5.11.1
python-jenkins==1.8.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.0